move:Dialogコンポーネントをdialogフォルダにまとめる
