+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import styles from "../styles/Home.module.css";
 import { useEffect, useState } from "react";
 import { formatInTimeZone } from "date-fns-tz";
 import Link from "next/link";
-import { Dialog } from "../components/Dialog";
+import { Dialog } from "../components/dialog/Dialog";
 import { Button } from "../components/button/Button";
 
 export type Todo = {
